+++ ceres-solver-src/CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
---
+- Include Eigen3 header files.
+- Include miniglog header files.
 
 ### Changed